DB: Do not use id 0 in database
This was inducing a big bug on the story that had the historyItem with
id 0 (if there was one)! when visiting a new element that goes in the
story and triggers the button, it erased the story and put only this
last element in it. So it looked like the db from lightyear and from
background were unsynced (until you refresh indexeddb in bg to see that
it had actually changed).