Refactor: fix getContent
we need to make sure that visit is updated once after paragraphs is
set, image is set and id is set so that the getContent is ended by the
background_controller.
Plus Items in lightyear need to be stored to see which one was
performing the getContent and refresh it.