Pool: use merge function for the pool.
That save the problem of the issue #144. Bag of words is only computed once at
create visit. But during the update, the putUnique Method of the pool make a remove
and replace so we losted all the infos about bagOfWords.
I also changed merged function to avoid problem if you don't put elements in the
right order. (safer)