UI: separate quotes from the same original paragraph
there was a bug here: if several quotes were found in the same
paragraph, first quote was created in a new <p>, then this <p> was
emptied (still present, but empty) and a new one was created with both
the 1st and second quote in it. etc. It could be valid to choose to
have quotes from the same original paragraph in a single <p> (with […]
in the middle for instance) but it is not what had been implemented, so
currently it is just a fix of the current implementation.